Optimise riffle_shuffle usage
Move looping inside of the riffle_shuffle function so that it can
allocate only a single temporary vector rather than allocating a new
vector on each iteration.  To reflect new argument, rename the
function to riffle_shuffle_n.

While at it, add unit tests for the function and for the observation
that encode+decode iterations bring canonical address back to its
original value.