orchestrate: zero call_depth when creating a sub VM
Because a) State is shared between all VM’s within a single execution
chain and b) call depth is tracked inside of State, recursive queries
(for example) count towards the call depth.  This is however not
desired since intention of call_depth is to track calls into the WASM
code only within a single VM.

Move the call_depth field out of State directly to Context which gets
reinitialised each time a sub-VM is created.