vm: refactor cosmwasm_system_run_hook function
Split handling of the Ok run result to a separate function.  This
makes cosmwasm_system_run_hook shorter and thus easier to read.
Furthermore, replace try_fold call by a loop.  This further flattens
the control flow graph making the code easier to grok.  As a result of
those changes, this results in negative delta which is always nice.